requirements.txt
# Main dependencies
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3

# Enhanced system dependencies
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
loguru>=0.7.2

# ML/AI (optional)
openai==1.3.0
transformers==4.35.2
torch==2.1.0
scikit-learn==1.3.2

# Utilities
pytz==2023.3.post1